requirements.txt
python-dotenv
requests
httpx
beautifulsoup4
feedparser
python-dateutil
lxml
pytz
python-slugify
google-generativeai
APSchedulertmdbv3api